Initial commit
Running from main in SquishyAquarium.java opens a window with a new squishy
aquarium. The aquarium is populated with squishes-- creatures composed
of nodes and springs that have generated shapes, properties, and
behaviors. Currently only a first-generation population is present, with
each squish's attributes determined at random. The first squish in the
population can be interacted with via mouse. Pressing any key refreshes
the population with newly generated squishes.