Rework serialization to improve readability and support subtree encoding
The subtree of any node n is N + S, where N is the set of descendant nodes in the bone-connected tree rooted at n (inclusive), and S is the set of springs s whose endpoints s.a, s.b ∈ N. A human-readable encoding of a node's subtree can be retrieved by invoking its stringifySubtree() method. This method will also accept a boolean includeStroke flag to determine whether the subtree's stroke (the list of actions involving members of the subtree-- an order-preserved subset of the entire body's stroke actions) is included in the output. includeStroke defaults to false when no arguments are provided.

SquishyBody provides a stringify() method that will return the subtree encoding of the head node, its stroke, and the set of constant values unique to the creature.

 These encodings are intended to be useful for copying and splicing creatures, two important processes for the evolutionary optimization algorithm soon to be implemented.