Rework the Stroke class to contain three separate action sequences
Updates were also made to dependent classes to accommodate this change. All three sequences are reflected when a squishyBody is encoded, and can be reconstructed by an instance of TreeParser.

The idea is that this will allow us to discriminate between sequences that result in different types of movement: specifically translation, positive Z-rotation, and negative Z-rotation. These movements, if sufficiently developed, should afford enough control to steer the squishyBody toward an arbitrary target (tests indicated that reversing the sequence of state changes is not always sufficient to reverse the direction of rotation).